[v0.0.1] TerraNetes: a HA k8s cluster spawner
Intended for use on AWS. The 0.0.x series should embed the last code
optimizations and a complete version of the docs + our
CONTRIBUTING.md.

Once we're done with that, we'll release version 0.1.0 and start our
CHANGELOG.md from there.

Sorry this all comes in one commit and therefore hides the some of the
logic behind this design implementations but it comes from a private
Git repo of ours whose history might have contained hard-coded private
credentials. For the record, most of this code has been produced
between August 2016 and October 2016 and we've been running it in our
production clusters since that date. Only one major issue occured when
migrating from etcd2 to etcd3 (a "split brain" in the flannel overlay
network after quorum was accidentaly lost on our etcd cluster...). So
it's already rather stable and we'll keep on improving it constantly.

For now, etcd and the Kubernetes master cluster need to be scaled
manually, but this will change soon.

Quick note about the License

All the code produced in this repo will be licensed under the CeCILL-C
license which is a transcription of the GNU GPLv3 in French law. The
text is available at the root of the repository, both in French and
English. In the US (and most other countries), it is strictly
equivalent to the GNU GPLv3 license.

Modifications to the source code should be made public. However, using
the code as "a library" (i.e. using its Terraform module) in
proprietary software (i.e. infrastructures) is obviously allowed by the
copyright holders.