Make etcd cluster more robust
This fixes issues in the implementation of the systemd chaining logic:

* The bootstrap relies on the same process but is a bit faster, we
benchmarked the systemd boot times with systemd-analyze to sleep a more
reasonable amount of time to ensure that the cloud init config has been
pulled before we attempt to SSH on the nodes etcd nodes to setup the
initial cluster.

* The boostraping is more segmented, more rigourous and therefore
doesn't sometimes fail at random. At the end of the bootstrapping
process, a key is written to the etcd (v2) store. That keys existence
conditons the status of the ELB health check, which has two virtuous
advantages:
  - etcd nodes added to the ELB only once the bootstrap/member join is
  complete. The health check can be trusted (and shouldn't destroy the
  cluster before it loses its majority... but will when it happens
  so... for now). In particular, in terraform we can leverage that to
  wait for the etcd cluster to be ready before spawning the Kubernetes
  master in a cleaner way.
  - These health checks can be used to destroy unhealthy nodes, for
  real.

* Same goes for the etcd cluster reconfiguration