[etcd] Refactor bootstraping process
It is much more rigorous now (no more sleep, fewer code duplication,
etc.). In addition, the use of the health check key in etcd highly
reduces the risk of "split brain" when talking to ETCD via the ELB.

We just need to finish to add tests and the etcd cluster will be ready
for general admission.